Changed Product entity fields to Object types
Having fields as Object types lets me distinguish when a field is not
entered. Rather than getting default values for primitives.
- For example, the default value for int is 0. However, I want to
  receive null if the values are not entered to handle those cases.

+++ src/main/java/com/abranlezama/ecommerceservice/model/Product.java
@@ -12,6 +12,6 @@ public class Product {
     @Column(length = 100)
     private String description;
     @Column(nullable = false)
-    private double price;
-    private int stock_quantity;
+    private Double price;
+    private Double stock_quantity;
 }
